Replace ecto_sql migrations with custom behaviour
Migrations in ecto_sql were understandably implemented with SQL-based
databases in mind. We got them to work in FDB, but ran into some
roadblocks in the use of a tenant handle as a prefix and some challenges
in using an FDB transaction to manage the migration step.

In this commit, we've removed the ecto_sql dep and replaced migrations
with a custom implementation. This comes with gotchas and caveats that
have been documented.